privacy functionality implemented

+++ src/app/api/blogs/[blogId]/route.ts
@@ -2,6 +2,8 @@ import { NextResponse } from "next/server";
 
 import Blog from "../../../../../models/blogModel";
 import connectMongoDB from "../../../../../lib/dbConnect";
+import { getUser } from "../../../../../utils/getUser";
+import AppError from "../../../../../utils/appError";
 
 // type for params
 type BlogParams = {
@@ -13,6 +15,9 @@ type BlogParams = {
 // Getting a particular blog
 export async function GET(req: Request, { params: { blogId } }: BlogParams) {
   connectMongoDB();
+  const user = await getUser();
+  if (!user) return NextResponse.json(new AppError(401, "Please login first"));
+
   const blog = await Blog.findById(blogId);
 
   return NextResponse.json({ status: "success", blog });
@@ -21,6 +26,9 @@ export async function GET(req: Request, { params: { blogId } }: BlogParams) {
 // Deleting a particular blog
 export async function DELETE(req: Request, { params: { blogId } }: BlogParams) {
   connectMongoDB();
+  const user = await getUser();
+  if (!user) return NextResponse.json(new AppError(401, "Please login first"));
+
   await Blog.findByIdAndDelete(blogId);
 
   return NextResponse.json({ status: "success", message: "Blog deleted" });
@@ -29,6 +37,8 @@ export async function DELETE(req: Request, { params: { blogId } }: BlogParams) {
 // Updating a particular blog
 export async function PATCH(req: Request, { params: { blogId } }: BlogParams) {
   connectMongoDB();
+  const user = await getUser();
+  if (!user) return NextResponse.json(new AppError(401, "Please login first"));
 
   const { title, content, likes, isGlobal } = await req.json();
   const blog = await Blog.findByIdAndUpdate(blogId, {
